+++ Gruntfile.js
@@ -28,7 +28,7 @@ module.exports = function( grunt ) {
             server: {
                 options: {
                     port: 8005,
-                    base: [ 'test/forms', 'build' ]
+                    base: [ 'test/forms', 'test/temp', 'build' ]
                 }
             },
             test: {