+++ CHANGELOG.md
@@ -7,7 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ---------------------
 ##### Fixed
 - Timepicker and Datetime Picker issues around empty and default values.
-- When loading a record with nested repeats the second+ series gets inserted out-of-order.
+- When loading a record with nested repeats the second+ series gets inserted out-of-order in the model.
+- When loading a record with nested repeats and additional nested groups, the nested repeats are not cloned in the view.
 
 [4.34.0] - 2017-05-22
 ---------------------